requirements.txt
python-telegram-bot==20.0a6
moviepy
openai-whisper
deep-translator
flask
